Move helpers to utils
