+++ frontend/src/metabase/metadata/pages/DataModel/components/PreviewSection/index.ts
@@ -1,1 +1,2 @@
 export * from "./PreviewSection";
+export { usePreviewType } from "./utils";
